requirements.txt
python-ternary>=1.0.8
numpy>=1.24.1
scipy>=1.10.0
PyQt5==5.15.11
pyqtgraph>=0.13.7